ID-list feature support
Reviewed By: czxttkl

Differential Revision: D29880479

fbshipit-source-id: 61c241d5570c7b81567974c50068a672e6058278

+++ reagent/preprocessing/sparse_preprocessor.py
@@ -102,6 +102,6 @@ class SparsePreprocessor(torch.nn.Module):
                 ret[self.id2name[fid]] = (
                     offsets.to(self.device),
                     idx_keys.to(self.device),
-                    weights.to(self.device),
+                    weights.to(self.device).float(),
                 )
         return ret
